Should have integrated the client-sever code into the main loop
